refactor(cli): extract sensor output topology
